requirements.txt
extra_streamlit_components==0.1.56
matplotlib==3.6.2
memory_profiler==0.61.0
numpy==1.23.4
pandas==1.5.1
pandas_profiling==3.6.2
Pillow==9.4.0
plotly==4.9.0
scikit_learn==1.2.0
seaborn==0.12.2
seaborn_analyzer==0.2.13
SQLAlchemy==1.4.46
streamlit==1.15.0
streamlit_menu==1.0.9
streamlit_option_menu==0.3.2
streamlit_pandas_profiling==0.1.3
